requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
openai==1.30.1
requests==2.32.2
openenv-core